renamed GeoUtils to Shapes
